+++ .gitignore
@@ -22,7 +22,10 @@ data/census/pums/data_dict_cache/
 data/eia/861/electric_utility_stats/
 data/eia/861/parquet/
 data/eia/heating_fuel_prices/parquet/
-data/nrel/
+data/resstock/utility/csv/
+data/resstock/utility/parquet/
+data/resstock/utility/shapefiles/
+data/resstock/utility/zips/
 data/eia/hourly_loads/zone_parquet/
 data/eia/hourly_loads/utility_parquet/
 data/fred/cpi/parquet/
