+++ data/resstock/load_curve/adjust_mf_electricity.py
