+++ data/resstock/load_curve/approximate_non_hp_load.py
