+++ data/resstock/metadata/add_vulnerability_columns.py
