+++ data/resstock/metadata/identify_hp_customers.py
