+++ data/resstock/metadata/identify_natgas_connection.py
