+++ tests/test_add_vulnerability_columns.py
@@ -8,7 +8,7 @@ from typing import cast
 import polars as pl
 import pytest
 
-from data.resstock.add_vulnerability_columns import (
+from data.resstock.metadata.add_vulnerability_columns import (
     PUMS_DISABLED_COLUMN,
     PUMS_PUMA_COLUMN,
     RESSTOCK_PUMA_COLUMN,
