+++ tests/test_identify_heating_type.py
@@ -2,7 +2,7 @@ from typing import cast
 
 import polars as pl
 
-from data.resstock.identify_heating_type import (
+from data.resstock.metadata.identify_heating_type import (
     IN_HVAC_COLUMN,
     UPGRADE_HVAC_COLUMN,
     identify_heating_type,
